+++ changelog.md
@@ -11,8 +11,9 @@
     - Can also be used as `uvx zipapps ...`
     - Or `zipapps.exe ...` on Windows, or `zipapps` on Unix
   - Added `--uv` option to accept a path to uv, speeding up `pip install`
+    - `--uv=uv` or `--uv=path/to/uv`
   - Added `--freeze-deps` as an alias for `--freeze-reqs`
-  - fix Deprecated since 3.10 `zipimporter.load_module`
+  - fix deprecated warning since 3.10 `zipimporter.load_module`
 - 2024.08.07
   - [**Compatible WARNING**]: update `sys_paths` insert index from `-1` to `0`
   - disable `--download-python`, use `python -m zipapps.download_python` instead